doc-serve-server/pyproject.toml
[tool.poetry]
name = "doc-serve-server"
version = "1.0.0"
description = "RAG-based document indexing and query REST API service"
authors = ["Spillwave Team"]
readme = "README.md"
packages = [{include = "doc_serve_server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
llama-index-core = "^0.14.0"
llama-index-readers-file = "^0.5.0"
llama-index-embeddings-openai = "^0.5.0"
llama-index-retrievers-bm25 = "^0.6.0"
rank-bm25 = "^0.2.2"
chromadb = "^0.5.0"
openai = "^1.57.0"
anthropic = "^0.40.0"
tiktoken = "^0.8.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.0"
click = "^8.1.0"
llama-index-llms-openai = "^0.6.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[tool.poetry.scripts]
doc-serve = "doc_serve_server.api.main:cli"
doc-serve-server = "doc_serve_server.api.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::pydantic._internal._generate_schema.UnsupportedFieldAttributeWarning",
]
